feat: [AD-008] components use contexts
